Infographics updated for:
Intubation guide

ALS and BLS modifications
ALS and BLS modifications also changed to a PDF viewer as was previously an image viewer.

Only slight issue is PDF viewer was unable to accept mixture of images in portrait or landscape. It seems to automatically present image in portrait.
I had a look at dev package online and there was no obvious fixes.
The images in landscape could be improved in terms of aspect, however they will suffice for current version.